Escaped special characters in the new mispellings
Escaped the '***' spam and 'alt+f4' spam because they broke regexp. Also
added a new mispellign for right and a forgotten comma.

Just a suggestion, cba to do it myself:
How about adding automatic escaping of the blocked words before adding
them to the regexp?
I can see a reason not to, that we can use regexp to block words, but so
far no such pattern is actually used.