ALLCAPS converted via CSS now, updated misty spam
I added a "stylers" class of options (but left them under the rewrite
list of options in the menu). The init code for the stylers probably
isn't as clean as it could be, but I wanted it to be short and similar
to the other options.

Also the misty spammers were starting to get clever by using unicode
characters, I added a few of them to the filtered list.