Detect which version of the chat we are using
The new chat filter doesn't work for the old chat. Ideally we would have
one single filter for both, but for now we are going to have to live
with this hacky solution.

We have to test for the new chat filter with a DOM element, and I don't
have the new chat, so I picked button.viewers. Not sure if that is the
best choice but that's one of the few unique things I see in the new
filter.