Stop waiting for the chat spinner
Now that we are overriding the Chat prototype instead of the chat objects
themselves we can get rid of the chat spinner wait. This makes the script
seem more responsive and also simplifies the code a bit (one less thing to
worry about when it comes to new chat compatibility)

The chat_is_loaded variables was kind of pointless too. We can't click on
any of the configuration buttons before the loading spinner goes away.