Add BetterTTV support to beta chat
BetterTTV takes over the chat and uses their own event handler for new
messages. If we want to support BetterTTV users we have to test for the
presence of BTTV and override that function. Fixes #76 for the beta
chat, and we'll just say that BTTV support for the old chat is no longer
a thing.