Stop filtering RIOT messages as emulator commands
Fixes issue #128, an issue introduced with PR #109.

Now the command filter only considers non-letter characters to be
part of a command if they look like separators: they must be adjacent to the
letters (no spaces) and must be at most 2 characters in length.

This should make the filter less general while still covering future changes to the emulator command syntax.